refactor : 리뷰 변경사항 반영(validlation, 양방향 로직, 패키징)
